+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 + [x] Input: remove metro style from input without role `input`. To add metro style to input without role `input`, use class `metro-input`.
 + [x] Textarea: remove metro style from input without role `input`. To add metro style to input without role `input`, use class `metro-input`.
 + [x] Utils Css: remove classes `.neb`, `.neb2`, `.h-center`, `.v-center`
++ [x] Input: add event `onAutocompleteSelect`. You can define this event with an attribute `data-on-autocomplete-select`. The event receives a one argument - selected value. issue #1615.
 
 ### 4.3.10
 + [x] Input: add API method `setAutocompleteList(array|string)`, issue #1576